+++ Source/TacticsModeRedux.Achtung/globalusings.cs
@@ -1,0 +1,4 @@
+global using System;
+global using HarmonyLib;
+global using Verse;
+global using RimWorld;