+++ Source/TacticsModeRedux/Core/Settings.cs
@@ -6,10 +6,13 @@ namespace TacticsModeRedux;
 [HotSwappable]
 public partial class Settings : ModSettings
 {
+    internal static bool hasAchtung = false;
+
     internal static bool _printDevMessages = false;
     internal static int _tacticalPauseTicks = 125;
     internal static bool _moveCameraOnPause = true;
     internal static bool _showMessageOnPause = true;
+    internal static bool _achtungForcedWorkPreventsPause = true;
 
     public override void ExposeData()
     {
@@ -20,6 +23,7 @@ public partial class Settings : ModSettings
         Scribe_Collections.Look(ref _alwaysPauseJobs, "alwaysPauseJobs", LookMode.Value);
         Scribe_Values.Look(ref _moveCameraOnPause, "moveCameraOnPause", true);
         Scribe_Values.Look(ref _showMessageOnPause, "showMessageOnPause", true);
+        Scribe_Values.Look(ref _achtungForcedWorkPreventsPause, "achtungForcedWorkPreventsPause", true);
     }
 
     private static Vector2 scrollPosition;
@@ -48,6 +52,14 @@ public partial class Settings : ModSettings
             ref _showMessageOnPause,
             "TM.ShowMessageOnPauseTooltip".Translate());
         listingStandard.Gap(4);
+
+        if (hasAchtung)
+        {
+            listingStandard.CheckboxLabeled(
+                "TM.AchtungForcedWorkPreventsPauseLabel".Translate(),
+                ref _achtungForcedWorkPreventsPause);
+            listingStandard.Gap(4);
+        }
 
         listingStandard.Label(_tacticalPauseTicks > 124
             ? "TM.TimeToPauseAfter".Translate(_tacticalPauseTicks.ToStringTicksToPeriod(allowSeconds: false))