+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2024-06-19
+
 ### Added
 
 -   Added not pausing when a colonist is performing Achtung!'s forced work. This behavior can be turned off in the settings.
@@ -34,7 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Added settings for which jobs to always pause after, as well as how long of a time to wait for after any job before pausing.
 
-[Unreleased]: https://github.com/ilyvion/TacticsModeRedux/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/ilyvion/TacticsModeRedux/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/ilyvion/TacticsModeRedux/releases/tag/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/ilyvion/TacticsModeRedux/releases/tag/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/ilyvion/TacticsModeRedux/releases/tag/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/ilyvion/TacticsModeRedux/releases/tag/v0.1.0